It turns out we don't want all this dll-decl stuff for mingw. So make the #ifdef correct for mscv-but-not-mingw.
I think the right fix would be to take all this logic out and
insert it via a MOE directive, just for the windows versions
of these files.  This is what ctemplate does.  But it's a
more-major change, and I'm looking to get this release out
today, so I'm submitting this (TBR), since it's a much more
minor change.


Revision created by MOE tool push_codebase.
MOE_MIGRATION=2837


git-svn-id: http://google-gflags.googlecode.com/svn/trunk@55 990008b9-e524-0410-ae84-45f8ed876473